Project: split VRDevice, ModelViewProj et PhysicsEngine files, move stereo stuff to Renderer, cleanup project files
