+++ cmake/CMakeLists_gl-win-x86.txt
@@ -245,7 +245,6 @@ add_executable(vpinball WIN32
    Material.h
    memutil.cpp
    memutil.h
-   mesh.cpp
    mesh.h
    mixer.cpp
    mixer.h
@@ -295,6 +294,8 @@ add_executable(vpinball WIN32
    src/core/Settings.h
    src/core/TableDB.cpp
    src/core/TableDB.h
+
+   src/math/ModelViewProj.h
 
    src/meshes/ballMesh.h
    src/meshes/bulbLightMesh.h
@@ -384,6 +385,8 @@ add_executable(vpinball WIN32
    src/physics/hittimer.h
    src/physics/kdtree.cpp
    src/physics/kdtree.h
+   src/physics/PhysicsEngine.cpp
+   src/physics/PhysicsEngine.h
    src/physics/quadtree.cpp
    src/physics/quadtree.h
 
@@ -422,6 +425,8 @@ add_executable(vpinball WIN32
    src/renderer/VertexBuffer.h
    src/renderer/ViewSetup.cpp
    src/renderer/ViewSetup.h
+   src/renderer/VRDevice.cpp
+   src/renderer/VRDevice.h
 
    vpinball.rc
    vpinball.idl
