+++ cmake/CMakeLists_win-x86.txt
@@ -280,7 +280,6 @@ add_executable(vpinball WIN32
    Material.h
    memutil.cpp
    memutil.h
-   mesh.cpp
    mesh.h
    mixer.cpp
    mixer.h
@@ -330,6 +329,8 @@ add_executable(vpinball WIN32
    src/core/Settings.h
    src/core/TableDB.cpp
    src/core/TableDB.h
+
+   src/math/ModelViewProj.h
 
    src/meshes/ballMesh.h
    src/meshes/bulbLightMesh.h
@@ -419,6 +420,8 @@ add_executable(vpinball WIN32
    src/physics/hittimer.h
    src/physics/kdtree.cpp
    src/physics/kdtree.h
+   src/physics/PhysicsEngine.cpp
+   src/physics/PhysicsEngine.h
    src/physics/quadtree.cpp
    src/physics/quadtree.h
 
@@ -457,6 +460,8 @@ add_executable(vpinball WIN32
    src/renderer/VertexBuffer.h
    src/renderer/ViewSetup.cpp
    src/renderer/ViewSetup.h
+   src/renderer/VRDevice.cpp
+   src/renderer/VRDevice.h
 
    vpinball.rc
    vpinball.idl
