+++ src/physics/PhysicsEngine.h
@@ -1,0 +1,7 @@
+#pragma once
+
+class PhysicsEngine final
+{
+public:
+
+};
